use chr field everywhere
